requirements.txt
fastapi
pydantic
google-generativeai
python-dotenv
PyPDF2